Fix command output
Removed a readline() that was causing it to skip the first line of the output. Removed a redundant try/except. The problem with omitting lines before the program exits uncleanly has been resolved to a c++ bug where the stdout/err buffers are not flushed before exiting when an assert() macro statement is encountered (be sure to flush these before any assert() statement in c++!). To ensure output is printed in real-time when using AnyCmd, it is recommended that you write something like these two lines in your respective language to disable output buffering, or else to simply flush the output buffer when needed: std::setbuf(stdout, nullptr);  std::setbuf(stderr, nullptr);